move publisher to shared module
