+++ fbw-common/src/systems/instruments/src/popup/index.ts
@@ -1,2 +1,1 @@
 export * from './popup';
-export * from './PopupSimvarPublisher';
