+++ fbw-common/src/systems/shared/src/publishers/Extras/PopupSimvarPublisher.tsx
@@ -21,3 +21,14 @@ export class PopupSimvarPublisher extends SimVarPublisher<PopupSimvars> {
     super(PopupSimvarPublisher.simvars, bus);
   }
 }
+
+export interface TodPauseOverlayState {
+  visible: boolean;
+  title: string;
+  message: string;
+}
+
+/** Controls visibility and content of the EFB pause-at-TOD overlay. */
+export interface TodPauseOverlayControlEvents {
+  tod_pause_overlay: TodPauseOverlayState;
+}
