+++ fbw-common/src/systems/shared/src/publishers/Extras/index.ts
@@ -1,2 +1,3 @@
 export * from './ExtrasSimVarPublisher';
 export * from './GsxSimVarPublisher';
+export * from './PopupSimvarPublisher';
